pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "piper_teleop"
version = "1.0.0"
description = "Accurate and High Throughput Remote Teleoperation Control System"
authors = [
    {name = "Zeno Hamers", email = "zeno@tactilerobotics.ai"},
    {name = "Alex Spiridonov", email = "alex@tactilerobotics.ai"},
]
keywords = ["robotics", "teleoperation", "vr", "robot-control"]

dependencies = [
    "opencv-python",
    "pyrealsense2",
    "pygame>=2.5",
    "pyaml",
    "python-dotenv",
    "meshcat",
    "piper_sdk==0.3.3",
    "lerobot==0.4.0",
]

requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
]

[project.scripts]
robotserver = "piper_teleop.robot_server.main:main_cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"piper_teleop.web_server" = ["web-ui/**/*"]
"piper_teleop.robot_server" = ["camera_streaming/calibration/*"] 

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 120
